Changes around testcontainers: 1. move common code in IdMappingDownloadControllerIT and IdMappingMessageConsumerIT to base class 2. Move TestConfig.java to different package to avoid another instance of test container from async download rest 3. remove unneeded annotation @Containers and @TestContainers 4. Handle lifecycle of testcontainers manually
