+++ async-download-rest/src/test/java/org/uniprot/api/async/download/controller/UniParcDownloadControllerIT.java
@@ -44,7 +44,6 @@ import org.uniprot.api.async.download.model.job.DownloadJob;
 import org.uniprot.api.async.download.model.job.uniparc.UniParcDownloadJob;
 import org.uniprot.api.common.repository.solrstream.FacetTupleStreamTemplate;
 import org.uniprot.api.common.repository.stream.common.TupleStreamTemplate;
-import org.uniprot.api.idmapping.common.service.TestConfig;
 import org.uniprot.api.rest.controller.ControllerITUtils;
 import org.uniprot.api.rest.download.model.JobStatus;
 import org.uniprot.api.rest.output.UniProtMediaType;
@@ -67,7 +66,6 @@ import lombok.extern.slf4j.Slf4j;
 @WebMvcTest({UniParcDownloadController.class})
 @ContextConfiguration(
         classes = {
-            TestConfig.class,
             UniParcDataStoreTestConfig.class,
             UniProtKBDataStoreTestConfig.class,
             AsyncDownloadRestApp.class,
