+++ async-download-rest/src/test/java/org/uniprot/api/async/download/controller/UniRefDownloadControllerIT.java
@@ -40,7 +40,6 @@ import org.uniprot.api.async.download.model.job.DownloadJob;
 import org.uniprot.api.async.download.model.job.uniref.UniRefDownloadJob;
 import org.uniprot.api.common.repository.solrstream.FacetTupleStreamTemplate;
 import org.uniprot.api.common.repository.stream.common.TupleStreamTemplate;
-import org.uniprot.api.idmapping.common.service.TestConfig;
 import org.uniprot.api.rest.controller.ControllerITUtils;
 import org.uniprot.api.rest.download.model.JobStatus;
 import org.uniprot.api.rest.output.UniProtMediaType;
@@ -63,7 +62,6 @@ import lombok.extern.slf4j.Slf4j;
 @WebMvcTest({UniRefDownloadController.class})
 @ContextConfiguration(
         classes = {
-            TestConfig.class,
             UniRefDataStoreTestConfig.class,
             UniProtKBDataStoreTestConfig.class,
             AsyncDownloadRestApp.class,
