+++ async-download-rest/src/test/java/org/uniprot/api/async/download/messaging/consumer/uniparc/UniParcMessageConsumerIT.java
@@ -40,7 +40,6 @@ import org.uniprot.api.async.download.model.job.uniparc.UniParcDownloadJob;
 import org.uniprot.api.async.download.model.request.uniparc.UniParcDownloadRequest;
 import org.uniprot.api.common.repository.solrstream.FacetTupleStreamTemplate;
 import org.uniprot.api.common.repository.stream.common.TupleStreamTemplate;
-import org.uniprot.api.idmapping.common.service.TestConfig;
 import org.uniprot.api.rest.controller.ControllerITUtils;
 import org.uniprot.api.rest.download.model.JobStatus;
 import org.uniprot.api.rest.output.context.FileType;
@@ -58,7 +57,6 @@ import com.jayway.jsonpath.JsonPath;
 @ActiveProfiles(profiles = {"offline", "idmapping"})
 @ContextConfiguration(
         classes = {
-            TestConfig.class,
             UniParcDataStoreTestConfig.class,
             UniProtKBDataStoreTestConfig.class,
             AsyncDownloadRestApp.class,
