+++ idmapping-common/src/test/java/org/uniprot/api/idmapping/common/service/IdMappingJobServiceTest.java
@@ -39,6 +39,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.web.client.RestClientException;
 import org.uniprot.api.common.exception.ResourceNotFoundException;
 import org.uniprot.api.idmapping.common.IdMappingDataStoreTestConfig;
+import org.uniprot.api.idmapping.common.TestConfig;
 import org.uniprot.api.idmapping.common.model.IdMappingJob;
 import org.uniprot.api.idmapping.common.model.IdMappingResult;
 import org.uniprot.api.idmapping.common.request.IdMappingJobRequest;
